all files split
