+++ srcs/parsing/utils/split_non_escaped.c
@@ -12,41 +12,6 @@
 
 #include "minishell.h"
 
-static short	is_whitespace(char c)
-{
-	if (c == ' ' || c == '\t')
-		return (1);
-	return (0);
-}
-
-static	short	is_delimiter(char input, char delimiter)
-{
-	if (delimiter == ' ' && is_whitespace(input))
-		return (1);
-	else if (input == delimiter)
-		return (1);
-	return (0);
-}
-
-static	int		count_args(char *input, char delimiter)
-{
-	int		i;
-	int		l;
-
-	l = 0;
-	i = 0;
-	while (input[i])
-	{
-		if (is_delimiter(input[i], delimiter) && input[i - 1] != '\\')
-		{
-			l++;
-			while (input[i] && !is_delimiter(input[i], delimiter))
-				i++;
-		}
-		i++;
-	}
-	return (l);
-}
 
 static	short	not_escaped(char *str, int i)
 {
