+++ srcs/search/search_utils.c
@@ -1,0 +1,24 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   search.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: migferna <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/09/12 10:08:20 by migferna          #+#    #+#             */
+/*   Updated: 2021/02/22 17:15:31 by migferna         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+void	to_lower_in(char *input)
+{
+	size_t	it;
+
+	it = -1;
+	while (input[++it])
+	{
+		input[it] = ft_tolower(input[it]);
+	}
+}
